GetNumCards and QueryCardInfo working
* Implement AiowdmInit()
* Implement GetNumCards()
* Implement QueryCardInfo()
* Fix device names to be lower case in accesio_pci_ids.h
  Comment out aiowdm.c

+++ include/accesio_pci_ioctl.h
@@ -14,7 +14,7 @@
 struct accesio_pci_card_descriptor
 {
   uint16_t device_id;
-  unsigned short int port_base;
+  size_t port_base;
   size_t port_size; //TODO: Better name?
   unsigned long name_size;
   char name[64];
